pyproject.toml
[project]
name = "krishimitra"
version = "0.1.0"
description = "Am ai assistant web application for helping the farmers"
authors = [
    {name = "Satyam Yadav (satyam.y@outlook.com)"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-community (>=0.3.24,<0.4.0)",
    "google-genai (>=1.18.0,<2.0.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "google-api-python-client (>=2.170.0,<3.0.0)",
    "google-cloud-aiplatform (>=1.95.1,<2.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "streamlit (>=1.45.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
